organised code, removed caching version 1, added comments
