+++ implementation/cache.cpp
@@ -1,50 +1,23 @@
 #include "cache.h"
+#include <string>
+#include <unordered_map>
+#include <unordered_set>
+#include <functional>
+#include <iostream>
 
-CacheValue::CacheValue()
-{
-}
+using namespace std;
 
-Cache::Cache(const set<string>::iterator &it)
-{
-    requests = 0;
-    hits = 0;
-    doc_words_begin=it;
-}
+bool WordCacheKey::operator==(const WordCacheKey& other) const{
+    return word == other.word &&
+            match_type == other.match_type &&
+            match_dist == other.match_dist;
+};
 
-CacheValue Cache::getHammingDistance(const string &query_word)
-{
-    requests++;
-    auto it = hamming_map.find(query_word);
-    if(it != hamming_map.end()){
-        hits++;
-        return it->second;
-    }
-    return CacheValue(DEFAULT_DISTANCE, doc_words_begin);
-}
+size_t std::hash<WordCacheKey>::operator()(const WordCacheKey& key) const {
 
-void Cache::addHammingDistance(const string &query_word, const unsigned int dist, const set<string>::iterator &it)
-{
-    hamming_map[query_word] = CacheValue(dist, it);
-}
-
-CacheValue Cache::getEditDistance(const string &query_word)
-{
-    requests++;
-    auto it = edit_map.find(query_word);
-    if(it != edit_map.end()){
-        hits++;
-        return it->second;
-    }
-    return CacheValue(DEFAULT_DISTANCE, doc_words_begin);
-}
-
-void Cache::addEditDistance(const string &query_word, const unsigned int dist, const set<string>::iterator &it)
-{
-    edit_map[query_word] = CacheValue(dist, it);
-}
-
-float Cache::hitRate()
-{
-    if(requests == 0) return 0;
-    return (float) hits / (float) requests;
+    size_t hash_value = std::hash<string>()(key.word);
+    hash_value ^= std::hash<int>()(key.match_type) + 0x9e3779b9 + (hash_value << 6) + (hash_value >> 2);
+    hash_value ^= std::hash<int>()(key.match_dist) + 0x9e3779b9 + (hash_value << 6) + (hash_value >> 2);
+    return hash_value;
+    
 }
