+++ include/cwrapper.h
@@ -3,20 +3,17 @@
 #include <sstream>
 #include <cache.h>
 #include <queries.h>
+#include <distance.h>
 #include <unordered_set>
 
 extern "C" {
 void* doc_str_to_doc_words(const char* doc_str);
-
-void* init_cache(void *document_words);
 
 void* init_ids();
 
 unsigned int queries_size(void *queries);
 
 void* query_by_index(void *queries, unsigned int index);
-
-bool match_query(void* queries, unsigned int index, void* document_words, void* cache);
 
 void add_ids(void* queries, unsigned int index, void* new_ids);
 
@@ -28,11 +25,11 @@ void end_query(void* queries, QueryID query_id);
 
 void* init_queries();
 
-bool match_query_og(void *queries, unsigned int index, void *document_words);
+bool match_query(void *queries, unsigned int index, void *document_words);
 
-bool matchQuery3(const Query& query, const set<string>& document_words);
+bool matchQuery(const Query& query, const set<string>& document_words);
 
-void* InitializeCache();
+void* initializeCache();
 
 void freeCache(void* cache);
