+++ python/test.py
@@ -6,7 +6,7 @@ import time
 from typing import TextIO
 from collections import deque
 from python.helper.helper import ErrorCode, MatchType
-# 0 -> core, 1 -> opt_core, 2 -> opt_core_v2, 3 -> opt_core_apache
+# 0 -> core, 1 -> opt_core, 2 -> dask
 def test_matching(which: int):
     INPUT_FILE_PATH = "test_data/small_test.txt"
     if which == 0:
@@ -16,9 +16,6 @@ def test_matching(which: int):
         from python.opt_core import DestroyIndex, EndQuery, GetNextAvailRes, InitializeIndex, MatchDocument, StartQuery
         OUTPUT_FILE_PATH = "results/result_opt_core.txt"
     elif which == 2:
-        from python.opt_core_v2 import DestroyIndex, EndQuery, GetNextAvailRes, InitializeIndex, MatchDocument, StartQuery
-        OUTPUT_FILE_PATH = "results/result_opt_core_v2.txt"
-    elif which == 3:
         from python.dask import DestroyIndex, EndQuery, GetNextAvailRes, InitializeIndex, MatchDocument, StartQuery
         OUTPUT_FILE_PATH = "results/result_dask.txt"
 
@@ -174,7 +171,3 @@ def test_matching(which: int):
     output_file.close()    
     #profiler.disable()
     #profiler.print_stats(sort="cumtime")
-
-
-if __name__ == "__main__":
-    test_matching(3)
